# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Agenda Electrónica",
        "description": "Documentación oficial para el proyecto de Ingeniería de Software UAC.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Inicio de sesión de usuario",
                "operationId": "78f0a66c9e72b510991be2d41660da95",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "test@agenda.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso"
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
        "/contacts": {
            "get": {
                "tags": [
                    "Contactos"
                ],
                "summary": "Listar contactos del usuario",
                "operationId": "7b8056a6c305d3995f6d8cc46d3cae65",
                "responses": {
                    "200": {
                        "description": "Lista de contactos obtenida"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Contactos"
                ],
                "summary": "Registrar un nuevo contacto",
                "operationId": "292652dc2e92a1f2c52aa7c864b8f718",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "correo"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "correo": {
                                        "type": "string",
                                        "example": "juan@example.com"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "9811234567"
                                    },
                                    "notas": {
                                        "type": "string",
                                        "example": "Compañero de la UAC"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contacto creado y correo enviado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/contacts/{id}": {
            "put": {
                "tags": [
                    "Contactos"
                ],
                "summary": "Actualizar un contacto existente",
                "operationId": "988ee956022c57234f5b5871a911543d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "correo": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contacto actualizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Contactos"
                ],
                "summary": "Eliminar un contacto",
                "operationId": "aa6804ea70c88459eaa3b88c8bab0195",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contacto eliminado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/events": {
            "get": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Listar eventos del calendario",
                "operationId": "3befe2e8582683a8dc70f7b706a5a97d",
                "responses": {
                    "200": {
                        "description": "Eventos recuperados"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Crear un evento y programar recordatorio",
                "operationId": "1ed79d9ab6d41b87387cad6237213e5d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "fecha_inicio"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "example": "Proyecto de Ingeniería"
                                    },
                                    "fecha_inicio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-03-03"
                                    },
                                    "hora": {
                                        "type": "string",
                                        "example": "10:00"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Entrega final con Fernando"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Evento programado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/events/{id}": {
            "delete": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Cancelar un evento",
                "operationId": "54be98df9381a8b89013d29e170a31fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Evento eliminado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/notifications/{id}/read": {
            "put": {
                "tags": [
                    "Notificaciones"
                ],
                "summary": "Marcar notificación como leída",
                "operationId": "8bc5203b95eaa5cf8614a66ced103386",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado actualizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/notifications/{id}": {
            "delete": {
                "tags": [
                    "Notificaciones"
                ],
                "summary": "Eliminar notificación de la lista",
                "operationId": "0f28a0400610f81f01e2d25ce4e0edff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notificación borrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Contactos",
            "description": "Operaciones para la libreta de direcciones"
        },
        {
            "name": "Eventos",
            "description": "Gestión de agenda y recordatorios"
        },
        {
            "name": "Notificaciones",
            "description": "Centro de alertas del sistema"
        },
        {
            "name": "Autenticación",
            "description": "Autenticación"
        }
    ]
}